Fixed record lookup error in Apprentice & Company edit forms
Previously records were retrieved from the database using
$_POST[‘inputName’]. This represented the value in the name field of
the edit form. This created an error where you couldn’t edit the name
of the Apprentice/Company because the table wouldn’t find a record with
the newly input name. The lookup function was refactored to use
params(0), which represents the name passed via the URL.